+++ mirar/pipelines/winter/models/_ref_stacks.py
@@ -4,7 +4,7 @@ Module to make reference stacks table
 from typing import ClassVar
 
 from pydantic import Field
-from sqlalchemy import VARCHAR, Column, Double, Float, Integer
+from sqlalchemy import VARCHAR, BigInteger, Column, Float, Integer
 
 from mirar.pipelines.winter.models.base_model import WinterBase, dec_field, ra_field
 from mirar.processors.sqldatabase.base_model import BaseDB
@@ -17,7 +17,7 @@ class RefStacksTable(WinterBase):
 
     __tablename__ = "refstacks"
 
-    stackid = Column(Double, primary_key=True, autoincrement=False)
+    stackid = Column(BigInteger, primary_key=True, autoincrement=False)
     ra_cent = Column(Float)
     dec_cent = Column(Float)
     ra0_0 = Column(Float)
